Move PersistedQueryDefaults into new file
