+++ index.js
@@ -1,6 +1,5 @@
 'use strict'
 
-const crypto = require('crypto')
 const fp = require('fastify-plugin')
 const LRU = require('tiny-lru')
 const routes = require('./lib/routes')
@@ -30,37 +29,9 @@ const buildGateway = require('./lib/gateway')
 const mq = require('mqemitter')
 const { PubSub } = require('./lib/subscriber')
 const { ErrorWithProps, FEDERATED_ERROR } = require('./lib/errors')
+const PersistedQueryDefaults = require('./lib/persistedQueryDefaults')
 
 const kLoaders = Symbol('fastify-gql.loaders')
-
-const PersistedQueryDefaults = {
-  Prepared: (persistedQueries) => ({
-    isPersistedQuery: (r) => r.persisted,
-    getHash: (r) => r.query,
-    getQueryFromHash: async (hash) => persistedQueries[hash]
-  }),
-  PreparedOnly: (persistedQueries) => ({
-    isPersistedQuery: (r) => true,
-    getHash: (r) => r.persisted ? r.query : false, // Only support persisted queries
-    getQueryFromHash: async (hash) => persistedQueries[hash]
-  }),
-  Automatic: () => {
-    // Initialize only in the scope of this server instance
-    const AUTOMATIC_PERSISTED_QUERIES = {}
-    return ({
-      isPersistedQuery: (r) => (r.extensions || {}).persistedQuery,
-      getHash: (r) => {
-        const { version, sha256Hash } = r.extensions.persistedQuery
-        return version === 1 ? sha256Hash : false
-      },
-      getQueryFromHash: async (hash) => AUTOMATIC_PERSISTED_QUERIES[hash],
-      getHashForQuery: (query) => crypto.createHash('sha256').update(query, 'utf8').digest('hex'),
-      saveQuery: async (hash, query) => { AUTOMATIC_PERSISTED_QUERIES[hash] = query },
-      notFoundError: 'PersistedQueryNotFound',
-      notSupportedError: 'PersistedQueryNotSupported'
-    })
-  }
-}
 
 function buildCache (opts) {
   if (Object.prototype.hasOwnProperty.call(opts, 'cache')) {
